Version 0.2alpha on new dev branch; remove use of Calc
Remove use of the Calc package for bignums, which was introduced in
SVN r4726 on 2018-08-14.  Use Emacs 27+ native bignums instead, which
simplifies the code quite a lot and should be faster.  The file esl.el
byte-compiles without errors but is not otherwise checked yet.  Also,
capitalise "Lisp" consistently and other minor updates.