Build Emacs REDUCE 3.8
Add current directory to load-path when running emacs.  Implement
non-recursive gcdn because the recursive version in "alg/numsup.red"
causes stack overflow in Emacs REDUCE 3.8 when building defint and
fps.  Don't build the support package, which is empty and just causes
an error when trying to load it.  Emacs REDUCE 3.8 now builds, but is
not yet tested.